+++ pytorch_lightning/trainer/trainer.py
@@ -51,6 +51,21 @@ def reduce_distributed_output(output, nb_gpus):
             output[k] = reduced
     return output
 
+import sys
+import pdb
+
+class ForkedPdb(pdb.Pdb):
+    """A Pdb subclass that may be used
+    from a forked multiprocessing child
+    """
+    def interaction(self, *args, **kwargs):
+        _stdin = sys.stdin
+        try:
+            sys.stdin = open('/dev/stdin')
+            pdb.Pdb.interaction(self, *args, **kwargs)
+        finally:
+            sys.stdin = _stdin
+
 
 class Trainer(TrainerIO):
 
@@ -1155,6 +1170,7 @@ class Trainer(TrainerIO):
         args = [batch, batch_nb]
         if len(self.optimizers) > 1:
             args.append(opt_idx)
+
         pdb.set_trace()
 
         if self.use_ddp:
